// src/redux/wishlist/wish.types.ts
import { CategoryItem } from "../shop/shop.types" 
export enum WISH_ACTION_TYPES {
  TOGGLE_WISH_LIST_HIDDEN = 'TTOGGLE_WISH_LIST_HIDDEN',
  ADD_ITEM_WISH = 'ADD_ITEM_WISH',
  REMOVE_ITEM_WISH = 'REMOVE_ITEM_WISH',
  // clear wish list
  CLEAR_ITEM_FROM_WISH_LIST = 'CLEAR_ITEM_FROM_WISH_LIST',
  CLEAR_WISH_LIST = 'CLEAR_WISH_LIST',
  // save wish list
  SAVE_WISH_LIST_START = 'SAVE_WISH_LIST_START',
  SAVE_WISH_LIST_SUCCESS = 'SAVE_WISH_LIST_SUCCESS',
  SAVE_WISH_LIST_FAILURE = 'SAVE_WISH_LIST_FAILURE',
  // delete wish list
  DELETE_WISH_LIST_START = 'DELETE_WISH_LIST_START',
  DELETE_WISH_LIST_SUCCESS = 'DELETE_WISH_LIST_SUCCESS',
  DELETE_WISH_LIST_FAILURE = 'DELETE_WISH_LIST_FAILURE'
};


export type WishlistItem =CategoryItem & {
  quantity:number
}